+++ CHANGELOG.md
@@ -11,3 +11,4 @@
 - Add Factory\Mapper\DoctrineORM\UserSettingMapperFactory test
 - Add Controller\Plugin\UserSettingPlugin test
 - Add View\Helper\UserSettingHelper test ( UserSettingHelper::getSetting() still partially tested, still can't test calling 'ZfcUserIdentity' view helper inside it with view property)
+- Add Service\UserSettingService test